# Conflicts:
#	TombEngine/Renderer/Renderer.h
#	TombEngine/Renderer/RendererDraw.cpp

+++ TombEngine/Renderer/ConstantBuffers/CameraMatrixBuffer.h
@@ -39,8 +39,8 @@ namespace TEN::Renderer::ConstantBuffers
 		//--
 		Vector4 FogColor;
 		//--
-		int FogMinDistance;
-		int FogMaxDistance;
+		float FogMinDistance;
+		float FogMaxDistance;
 		float NearPlane;
 		float FarPlane;
 		//--
